pyproject.toml
[tool.poetry]
name = "salary-count-bot"
version = "0.1.0"
description = "Bot to count salary"
authors = ["Kirill Danilkov <k.danilkov.work@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyTelegramBotAPI = "^4.12.0"
python-dotenv = "^1.0.0"
telebot-calendar = "^1.2"
ipython = "^8.14.0"
pytz = "^2023.3"
schedule = "^1.2.0"
sqlalchemy = "^2.0.20"
black = "^23.9.0"
jinja2 = "^3.1.2"
weasyprint = "^59.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
